Library Management System
This project is to develop a student report management system that can be
used to maintain the record of the student that is currently maintained in
majority of the database system to maintain the record of different students
in schools, colleges, and in various universities. The traditional system that
using by most of the schools, colleges and institutions is the traditional
manual marking of the attendence which means all works and procedures is
recorded through manpower manual work and it consist of a huge amount of
paperwork that is not effective. This causes the business to encounter trouble
regarding human error due to the huge amount of manpower manual work
that operating in each institution routine. Thus, this computerized student
the report management system is designed to assist the institution in terms of
having better management as well as easier to handle daily operation